refactor uart.h
