+++ board/main.c
@@ -3,7 +3,6 @@
 
 #include "drivers/pwm.h"
 #include "drivers/usb.h"
-#include "drivers/uart.h"
 #include "drivers/gmlan_alt.h"
 #include "drivers/kline_init.h"
